Structure and bug fixes.
