+++ queueservice/Dockerfile-local
@@ -6,14 +6,13 @@ ARG DEBIAN_FRONTEND=noninteractive
 ARG PROCESS_PORT=5000
 ENV PORT ${PROCESS_PORT}
 
-
 # Install dependencies
-COPY queue/requirements.txt /
+COPY queueservice/requirements.txt /
 RUN /docker_venv/bin/pip3 install -r /requirements.txt
 
 # Setup Working Directory
 RUN mkdir /queue
-ADD ./queue /queue
+ADD ./queueservice /queue
 WORKDIR /queue
 
 EXPOSE ${PORT}
