+++ renderservice/render/tests/resources/BaseResourceTest.py
