+++ renderservice/render/tests/resources/test_binary_to_alphabet.py
@@ -1,22 +1,22 @@
 import itertools
 from render.daemon.ResourceGenerator import TaskError
-from render.tests.BaseResourceTest import BaseResourceTest
+from render.tests.resources.BaseResourceTest import BaseResourceTest
 
 
-class ParityCardsResourceTest(BaseResourceTest):
+class BinaryToAlphabetResourceTest(BaseResourceTest):
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        self.module = "parity_cards"
-        self.BASE_URL = "resources/parity-cards.html"
+        self.module = "binary_to_alphabet"
+        self.BASE_URL = "resources/binary-to-alphabet.html"
         self.TASK_TEMPLATE = {
-            "resource_slug": "parity-cards",
-            "resource_name": "Parity Cards",
-            "resource_view": "parity_cards",
+            "resource_slug": "binary-to-alphabet",
+            "resource_name": "Binary To Alphabet",
+            "resource_view": "binary_to_alphabet",
             "url": None
         }
 
-    def test_parity_cards_small_resource_generation_valid_configurations(self):
+    def test_binary_to_alphabet_resource_generation_valid_configurations(self):
         resource_module = self.load_module()
         valid_options = resource_module.valid_options()
         valid_options["header_text"] = ["", "Example header"]
@@ -30,7 +30,7 @@ class ParityCardsResourceTest(BaseResourceTest):
         ]
 
         print()
-        print("Testing Parity Cards:")
+        print("Testing Binary To Alphabet:")
         for combination in combinations:
             print("   - Testing combination: {} ... ".format(combination), end="")
             url = self.BASE_URL + self.query_string(combination)
@@ -41,9 +41,9 @@ class ParityCardsResourceTest(BaseResourceTest):
             filename, pdf = self.generator.generate_resource_pdf(task)
             print("ok")
 
-    def test_parity_cards_resource_generation_missing_back_colour_parameter(self):
+    def test_binary_to_alphabet_resource_generation_missing_worksheet_version_parameter(self):
         combination = {
-            "paper_size": "a4",
+            "paper_size": "letter",
             "header_text": "",
             "copies": 1
         }
@@ -56,10 +56,11 @@ class ParityCardsResourceTest(BaseResourceTest):
         with self.assertRaises(TaskError):
             filename, pdf = self.generator.generate_resource_pdf(task)
 
-    def test_parity_cards_resource_generation_missing_paper_size_parameter(self):
+    def test_binary_to_alphabet_resource_generation_missing_paper_size_parameter(self):
         combination = {
-            "back_colour": "black",
+            "worksheet_version": "student",
             "header_text": "",
+            "copies": 1
         }
 
         url = self.BASE_URL + self.query_string(combination)
@@ -70,11 +71,11 @@ class ParityCardsResourceTest(BaseResourceTest):
         with self.assertRaises(TaskError):
             filename, pdf = self.generator.generate_resource_pdf(task)
 
-    def test_parity_cards_resource_generation_missing_header_text_parameter(self):
-        expected_filename = "Parity Cards (black back - a4).pdf"
+    def test_binary_to_alphabet_resource_generation_missing_header_text_parameter(self):
+        expected_filename = "Binary To Alphabet (teacher - letter).pdf"
         combination = {
-            "back_colour": "black",
-            "paper_size": "a4",
+            "worksheet_version": "teacher",
+            "paper_size": "letter",
             "copies": 1
         }
