+++ renderservice/render/tests/resources/test_left_right_cards.py
@@ -1,22 +1,22 @@
 import itertools
 from render.daemon.ResourceGenerator import TaskError
-from render.tests.BaseResourceTest import BaseResourceTest
+from render.tests.resources.BaseResourceTest import BaseResourceTest
 
 
-class BinaryToAlphabetResourceTest(BaseResourceTest):
+class LeftRightCardsResourceTest(BaseResourceTest):
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        self.module = "binary_to_alphabet"
-        self.BASE_URL = "resources/binary-to-alphabet.html"
+        self.module = "left_right_cards"
+        self.BASE_URL = "resources/left-right-cards.html"
         self.TASK_TEMPLATE = {
-            "resource_slug": "binary-to-alphabet",
-            "resource_name": "Binary To Alphabet",
-            "resource_view": "binary_to_alphabet",
+            "resource_slug": "left-right-cards",
+            "resource_name": "Left and Right Cards",
+            "resource_view": "left_right_cards",
             "url": None
         }
 
-    def test_binary_to_alphabet_resource_generation_valid_configurations(self):
+    def test_left_right_cards_resource_generation_valid_configurations(self):
         resource_module = self.load_module()
         valid_options = resource_module.valid_options()
         valid_options["header_text"] = ["", "Example header"]
@@ -30,7 +30,7 @@ class BinaryToAlphabetResourceTest(BaseResourceTest):
         ]
 
         print()
-        print("Testing Binary To Alphabet:")
+        print("Testing Left and Right Cards:")
         for combination in combinations:
             print("   - Testing combination: {} ... ".format(combination), end="")
             url = self.BASE_URL + self.query_string(combination)
@@ -41,10 +41,9 @@ class BinaryToAlphabetResourceTest(BaseResourceTest):
             filename, pdf = self.generator.generate_resource_pdf(task)
             print("ok")
 
-    def test_binary_to_alphabet_resource_generation_missing_worksheet_version_parameter(self):
+    def test_left_right_cards_resource_generation_missing_paper_size_parameter(self):
         combination = {
-            "paper_size": "letter",
-            "header_text": "",
+            "header_text": "Example header text",
             "copies": 1
         }
 
@@ -56,26 +55,10 @@ class BinaryToAlphabetResourceTest(BaseResourceTest):
         with self.assertRaises(TaskError):
             filename, pdf = self.generator.generate_resource_pdf(task)
 
-    def test_binary_to_alphabet_resource_generation_missing_paper_size_parameter(self):
+    def test_left_right_cards_resource_generation_missing_header_text_parameter(self):
+        expected_filename = "Left and Right Cards (a4).pdf"
         combination = {
-            "worksheet_version": "student",
-            "header_text": "",
-            "copies": 1
-        }
-
-        url = self.BASE_URL + self.query_string(combination)
-        task = self.TASK_TEMPLATE.copy()
-        task.update(combination)
-        task["url"] = url
-
-        with self.assertRaises(TaskError):
-            filename, pdf = self.generator.generate_resource_pdf(task)
-
-    def test_binary_to_alphabet_resource_generation_missing_header_text_parameter(self):
-        expected_filename = "Binary To Alphabet (teacher - letter).pdf"
-        combination = {
-            "worksheet_version": "teacher",
-            "paper_size": "letter",
+            "paper_size": "a4",
             "copies": 1
         }
