+++ renderservice/render/tests/resources/test_parity_cards.py
@@ -1,22 +1,22 @@
 import itertools
 from render.daemon.ResourceGenerator import TaskError
-from render.tests.BaseResourceTest import BaseResourceTest
+from render.tests.resources.BaseResourceTest import BaseResourceTest
 
 
-class PianoKeysResourceTest(BaseResourceTest):
+class ParityCardsResourceTest(BaseResourceTest):
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        self.module = "piano_keys"
-        self.BASE_URL = "resources/piano-keys.html"
+        self.module = "parity_cards"
+        self.BASE_URL = "resources/parity-cards.html"
         self.TASK_TEMPLATE = {
-            "resource_slug": "piano-keys",
-            "resource_name": "Piano Keys",
-            "resource_view": "piano_keys",
+            "resource_slug": "parity-cards",
+            "resource_name": "Parity Cards",
+            "resource_view": "parity_cards",
             "url": None
         }
 
-    def test_piano_keys_resource_generation_valid_configurations(self):
+    def test_parity_cards_small_resource_generation_valid_configurations(self):
         resource_module = self.load_module()
         valid_options = resource_module.valid_options()
         valid_options["header_text"] = ["", "Example header"]
@@ -30,7 +30,7 @@ class PianoKeysResourceTest(BaseResourceTest):
         ]
 
         print()
-        print("Testing Piano Keys:")
+        print("Testing Parity Cards:")
         for combination in combinations:
             print("   - Testing combination: {} ... ".format(combination), end="")
             url = self.BASE_URL + self.query_string(combination)
@@ -41,7 +41,7 @@ class PianoKeysResourceTest(BaseResourceTest):
             filename, pdf = self.generator.generate_resource_pdf(task)
             print("ok")
 
-    def test_piano_keys_resource_generation_missing_highlight_parameter(self):
+    def test_parity_cards_resource_generation_missing_back_colour_parameter(self):
         combination = {
             "paper_size": "a4",
             "header_text": "",
@@ -56,11 +56,10 @@ class PianoKeysResourceTest(BaseResourceTest):
         with self.assertRaises(TaskError):
             filename, pdf = self.generator.generate_resource_pdf(task)
 
-    def test_piano_keys_resource_generation_missing_paper_size_parameter(self):
+    def test_parity_cards_resource_generation_missing_paper_size_parameter(self):
         combination = {
-            "highlight": False,
+            "back_colour": "black",
             "header_text": "",
-            "copies": 1
         }
 
         url = self.BASE_URL + self.query_string(combination)
@@ -71,10 +70,10 @@ class PianoKeysResourceTest(BaseResourceTest):
         with self.assertRaises(TaskError):
             filename, pdf = self.generator.generate_resource_pdf(task)
 
-    def test_piano_keys_resource_generation_missing_header_text_parameter(self):
-        expected_filename = "Piano Keys (no highlight - a4).pdf"
+    def test_parity_cards_resource_generation_missing_header_text_parameter(self):
+        expected_filename = "Parity Cards (black back - a4).pdf"
         combination = {
-            "highlight": False,
+            "back_colour": "black",
             "paper_size": "a4",
             "copies": 1
         }
