+++ renderservice/render/tests/resources/test_piano_keys.py
@@ -1,22 +1,22 @@
 import itertools
 from render.daemon.ResourceGenerator import TaskError
-from render.tests.BaseResourceTest import BaseResourceTest
+from render.tests.resources.BaseResourceTest import BaseResourceTest
 
 
-class SortingNetworkResourceTest(BaseResourceTest):
+class PianoKeysResourceTest(BaseResourceTest):
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        self.module = "sorting_network"
-        self.BASE_URL = "resources/modulo-clock.html"
+        self.module = "piano_keys"
+        self.BASE_URL = "resources/piano-keys.html"
         self.TASK_TEMPLATE = {
-            "resource_slug": "sorting-network",
-            "resource_name": "Sorting Network",
-            "resource_view": "sorting_network",
+            "resource_slug": "piano-keys",
+            "resource_name": "Piano Keys",
+            "resource_view": "piano_keys",
             "url": None
         }
 
-    def test_sorting_network_resource_generation_valid_configurations(self):
+    def test_piano_keys_resource_generation_valid_configurations(self):
         resource_module = self.load_module()
         valid_options = resource_module.valid_options()
         valid_options["header_text"] = ["", "Example header"]
@@ -30,7 +30,7 @@ class SortingNetworkResourceTest(BaseResourceTest):
         ]
 
         print()
-        print("Testing Sorting Network:")
+        print("Testing Piano Keys:")
         for combination in combinations:
             print("   - Testing combination: {} ... ".format(combination), end="")
             url = self.BASE_URL + self.query_string(combination)
@@ -41,7 +41,7 @@ class SortingNetworkResourceTest(BaseResourceTest):
             filename, pdf = self.generator.generate_resource_pdf(task)
             print("ok")
 
-    def test_sorting_network_resource_generation_missing_prefilled_values_parameter(self):
+    def test_piano_keys_resource_generation_missing_highlight_parameter(self):
         combination = {
             "paper_size": "a4",
             "header_text": "",
@@ -56,9 +56,9 @@ class SortingNetworkResourceTest(BaseResourceTest):
         with self.assertRaises(TaskError):
             filename, pdf = self.generator.generate_resource_pdf(task)
 
-    def test_sorting_network_resource_generation_missing_paper_size_parameter(self):
+    def test_piano_keys_resource_generation_missing_paper_size_parameter(self):
         combination = {
-            "prefilled_values": "blank",
+            "highlight": False,
             "header_text": "",
             "copies": 1
         }
@@ -71,10 +71,10 @@ class SortingNetworkResourceTest(BaseResourceTest):
         with self.assertRaises(TaskError):
             filename, pdf = self.generator.generate_resource_pdf(task)
 
-    def test_sorting_network_generation_missing_header_text_parameter(self):
-        expected_filename = "Sorting Network (blank - a4).pdf"
+    def test_piano_keys_resource_generation_missing_header_text_parameter(self):
+        expected_filename = "Piano Keys (no highlight - a4).pdf"
         combination = {
-            "prefilled_values": "blank",
+            "highlight": False,
             "paper_size": "a4",
             "copies": 1
         }
