+++ renderservice/render/tests/resources/test_searching_cards.py
@@ -1,22 +1,22 @@
 import itertools
 from render.daemon.ResourceGenerator import TaskError
-from render.tests.BaseResourceTest import BaseResourceTest
+from render.tests.resources.BaseResourceTest import BaseResourceTest
 
 
-class BinaryCardsSmallResourceTest(BaseResourceTest):
+class SearchingCardsResourceTest(BaseResourceTest):
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        self.module = "binary_cards_small"
-        self.BASE_URL = "resources/binary-cards-small.html"
+        self.module = "searching_cards"
+        self.BASE_URL = "resources/searching-cards.html"
         self.TASK_TEMPLATE = {
-            "resource_slug": "binary-cards",
-            "resource_name": "Binary Cards (small)",
-            "resource_view": "binary_cards_small",
+            "resource_slug": "searching-cards",
+            "resource_name": "Searching Cards",
+            "resource_view": "searching_cards",
             "url": None
         }
 
-    def test_binary_cards_small_resource_generation_valid_configurations(self):
+    def test_searching_cards_resource_generation_valid_configurations(self):
         resource_module = self.load_module()
         valid_options = resource_module.valid_options()
         valid_options["header_text"] = ["", "Example header"]
@@ -30,7 +30,7 @@ class BinaryCardsSmallResourceTest(BaseResourceTest):
         ]
 
         print()
-        print("Testing Binary Cards (small):")
+        print("Testing Searching Cards:")
         for combination in combinations:
             print("   - Testing combination: {} ... ".format(combination), end="")
             url = self.BASE_URL + self.query_string(combination)
@@ -41,10 +41,10 @@ class BinaryCardsSmallResourceTest(BaseResourceTest):
             filename, pdf = self.generator.generate_resource_pdf(task)
             print("ok")
 
-    def test_binary_cards_small_resource_generation_missing_dot_count_parameter(self):
+    def test_searching_cards_resource_generation_missing_number_cards_parameter(self):
         combination = {
-            "number_bits": "4",
-            "black_back": True,
+            "max_number": "99",
+            "help_sheet": True,
             "paper_size": "a4",
             "header_text": "",
             "copies": 1
@@ -58,10 +58,10 @@ class BinaryCardsSmallResourceTest(BaseResourceTest):
         with self.assertRaises(TaskError):
             filename, pdf = self.generator.generate_resource_pdf(task)
 
-    def test_binary_cards_small_resource_generation_missing_black_back_parameter(self):
+    def test_searching_cards_resource_generation_missing_max_number_parameter(self):
         combination = {
-            "number_bits": "4",
-            "dot_counts": True,
+            "number_cards": "15",
+            "help_sheet": True,
             "paper_size": "a4",
             "header_text": "",
             "copies": 1
@@ -75,10 +75,10 @@ class BinaryCardsSmallResourceTest(BaseResourceTest):
         with self.assertRaises(TaskError):
             filename, pdf = self.generator.generate_resource_pdf(task)
 
-    def test_binary_cards_small_resource_generation_missing_number_bits_parameter(self):
+    def test_searching_cards_resource_generation_missing_help_sheet_parameter(self):
         combination = {
-            "dot_counts": True,
-            "black_back": True,
+            "number_cards": "15",
+            "max_number": "99",
             "paper_size": "a4",
             "header_text": "",
             "copies": 1
@@ -92,11 +92,11 @@ class BinaryCardsSmallResourceTest(BaseResourceTest):
         with self.assertRaises(TaskError):
             filename, pdf = self.generator.generate_resource_pdf(task)
 
-    def test_binary_cards_small_resource_generation_missing_paper_size_parameter(self):
+    def test_searching_cards_resource_generation_missing_paper_size_parameter(self):
         combination = {
-            "number_bits": "4",
-            "dot_counts": True,
-            "black_back": True,
+            "number_cards": "15",
+            "max_number": "99",
+            "help_sheet": True,
             "header_text": "",
             "copies": 1
         }
@@ -109,12 +109,12 @@ class BinaryCardsSmallResourceTest(BaseResourceTest):
         with self.assertRaises(TaskError):
             filename, pdf = self.generator.generate_resource_pdf(task)
 
-    def test_binary_cards_small_resource_generation_missing_header_text_parameter(self):
-        expected_filename = "Binary Cards (small) (4 bits - with dot counts - with black back - a4).pdf"
+    def test_searching_cards_resource_generation_missing_header_text_parameter(self):
+        expected_filename = "Searching Cards (15 cards - 0 to 99 - with helper sheet - a4).pdf"
         combination = {
-            "dot_counts": True,
-            "number_bits": "4",
-            "black_back": True,
+            "number_cards": "15",
+            "max_number": "99",
+            "help_sheet": True,
             "paper_size": "a4",
             "copies": 1
         }
@@ -126,20 +126,3 @@ class BinaryCardsSmallResourceTest(BaseResourceTest):
 
         filename, pdf = self.generator.generate_resource_pdf(task)
         self.assertEqual(filename, expected_filename)
-
-    def test_binary_cards_small_resource_generation_invalid_black_back_parameter(self):
-        combination = {
-            "number_bits": "4",
-            "dot_counts": True,
-            "black_back": "maybe",
-            "header_text": "",
-            "copies": 1
-        }
-
-        url = self.BASE_URL + self.query_string(combination)
-        task = self.TASK_TEMPLATE.copy()
-        task.update(combination)
-        task["url"] = url
-
-        with self.assertRaises(TaskError):
-            filename, pdf = self.generator.generate_resource_pdf(task)
