+++ renderservice/render/tests/resources/test_treasure_hunt.py
@@ -1,22 +1,22 @@
 import itertools
 from render.daemon.ResourceGenerator import TaskError
-from render.tests.BaseResourceTest import BaseResourceTest
+from render.tests.resources.BaseResourceTest import BaseResourceTest
 
 
-class BinaryWindowsResourceTest(BaseResourceTest):
+class TreasureHuntResourceTest(BaseResourceTest):
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        self.module = "binary_windows"
-        self.BASE_URL = "resources/binary-windows.html"
+        self.module = "treasure_hunt"
+        self.BASE_URL = "resources/treasure-hunt.html"
         self.TASK_TEMPLATE = {
-            "resource_slug": "binary-windows",
-            "resource_name": "Binary Windows",
-            "resource_view": "binary_windows",
+            "resource_slug": "treasure-hunt",
+            "resource_name": "Treasure Hunt",
+            "resource_view": "treasure_hunt",
             "url": None
         }
 
-    def test_binary_windows_resource_generation_valid_configurations(self):
+    def test_treasure_hunt_resource_generation_valid_configurations(self):
         resource_module = self.load_module()
         valid_options = resource_module.valid_options()
         valid_options["header_text"] = ["", "Example header"]
@@ -30,7 +30,7 @@ class BinaryWindowsResourceTest(BaseResourceTest):
         ]
 
         print()
-        print("Testing Binary Windows:")
+        print("Testing Treasure Hunt:")
         for combination in combinations:
             print("   - Testing combination: {} ... ".format(combination), end="")
             url = self.BASE_URL + self.query_string(combination)
@@ -41,12 +41,13 @@ class BinaryWindowsResourceTest(BaseResourceTest):
             filename, pdf = self.generator.generate_resource_pdf(task)
             print("ok")
 
-    def test_binary_windows_resource_generation_missing_dot_count_parameter(self):
+    def test_treasure_hunt_resource_generation_missing_prefilled_values_parameter(self):
         combination = {
-            "number_bits": "8",
-            "value_type": "binary",
+            "number_order": "sorted",
+            "instructions": True,
+            "art": "colour",
             "paper_size": "a4",
-            "header_text": "Example header text",
+            "header_text": "",
             "copies": 1
         }
 
@@ -58,12 +59,13 @@ class BinaryWindowsResourceTest(BaseResourceTest):
         with self.assertRaises(TaskError):
             filename, pdf = self.generator.generate_resource_pdf(task)
 
-    def test_binary_windows_resource_generation_missing_number_bits_parameter(self):
+    def test_treasure_hunt_resource_generation_missing_number_order_parameter(self):
         combination = {
-            "dot_counts": True,
-            "value_type": "binary",
+            "prefilled_values": "blank",
+            "instructions": True,
+            "art": "colour",
             "paper_size": "a4",
-            "header_text": "Example header text",
+            "header_text": "",
             "copies": 1
         }
 
@@ -75,12 +77,13 @@ class BinaryWindowsResourceTest(BaseResourceTest):
         with self.assertRaises(TaskError):
             filename, pdf = self.generator.generate_resource_pdf(task)
 
-    def test_binary_windows_resource_generation_missing_value_type_parameter(self):
+    def test_treasure_hunt_resource_generation_missing_instructions_parameter(self):
         combination = {
-            "dot_counts": True,
-            "number_bits": "8",
+            "number_order": "sorted",
+            "prefilled_values": "blank",
+            "art": "colour",
             "paper_size": "a4",
-            "header_text": "Example header text",
+            "header_text": "",
             "copies": 1
         }
 
@@ -92,12 +95,13 @@ class BinaryWindowsResourceTest(BaseResourceTest):
         with self.assertRaises(TaskError):
             filename, pdf = self.generator.generate_resource_pdf(task)
 
-    def test_binary_windows_resource_generation_missing_paper_size_parameter(self):
+    def test_treasure_hunt_resource_generation_missing_art_parameter(self):
         combination = {
-            "dot_counts": True,
-            "number_bits": "8",
-            "value_type": "binary",
-            "header_text": "Example header text",
+            "number_order": "sorted",
+            "prefilled_values": "blank",
+            "instructions": True,
+            "paper_size": "a4",
+            "header_text": "",
             "copies": 1
         }
 
@@ -109,12 +113,33 @@ class BinaryWindowsResourceTest(BaseResourceTest):
         with self.assertRaises(TaskError):
             filename, pdf = self.generator.generate_resource_pdf(task)
 
-    def test_binary_windows_resource_generation_missing_header_text_parameter(self):
-        expected_filename = "Binary Windows (8 bits - lightbulb - with dot counts).pdf"
+    def test_treasure_hunt_resource_generation_missing_paper_size_parameter(self):
         combination = {
-            "dot_counts": True,
-            "number_bits": "8",
-            "value_type": "lightbulb",
+            "number_order": "sorted",
+            "instructions": True,
+            "art": "colour",
+            "prefilled_values": "blank",
+            "number_order": "sorted",
+            "header_text": "",
+            "copies": 1
+        }
+
+        url = self.BASE_URL + self.query_string(combination)
+        task = self.TASK_TEMPLATE.copy()
+        task.update(combination)
+        task["url"] = url
+
+        with self.assertRaises(TaskError):
+            filename, pdf = self.generator.generate_resource_pdf(task)
+
+    def test_treasure_hunt_generation_missing_header_text_parameter(self):
+        expected_filename = "Treasure Hunt (blank - full colour - with instructions - a4).pdf"
+        combination = {
+            "number_order": "sorted",
+            "instructions": True,
+            "art": "colour",
+            "prefilled_values": "blank",
+            "number_order": "sorted",
             "paper_size": "a4",
             "copies": 1
         }
